pyproject.toml
[project]
name = "nonebot-plugin-apod"
version = "1.0.7"
description = "定时发送 NASA 每日提供的天文图片"
readme = "README.md"
authors = [{ name = "lyqgzbl", email = "admin@lyqgzbl.com" }]
license = "MIT"
keywords = ["nonebot2", "NASA"]
requires-python = ">=3.10"
dependencies = [
  "nonebot2 >=2.2.1, <3.0.0",
  "httpx >=0.28.0, <1.0.0",
  "nonebot-plugin-argot >=0.2.2",
  "nonebot-plugin-localstore >=0.7.0, <1.0.0",
  "nonebot-plugin-apscheduler >=0.5.0, <1.0.0",
  "nonebot-plugin-alconna >=0.57.1, <1.0.0",
  "nonebot-plugin-htmlrender >=0.5.0, <1.0.0",
]

[project.urls]
Homepage = "https://github.com/lyqgzbl/nonebot-plugin-apod"
Repository = "https://github.com/lyqgzbl/nonebot-plugin-apod"
Documentation = "https://github.com/lyqgzbl/nonebot-plugin-apod#readme"

[build-system]
requires = ["uv_build >=0.8.3, <0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "nonebot_plugin_apod"
module-root = ""

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
